(BSR)[PRO] feat: split test
