+++ pro/cypress/e2e/collective/adage.cy.ts
